fixed buggy audio writing
svn path=/trunk/kdemultimedia/k3b/; revision=125948

+++ src/audio/k3baudioontheflyjob.cpp
@@ -34,6 +34,7 @@
 #include <qstring.h>
 #include <qtimer.h>
 #include <qdatetime.h>
+#include <qfile.h>
 
 #include <iostream>
 #include <cmath>
@@ -44,6 +45,7 @@ K3bAudioOnTheFlyJob::K3bAudioOnTheFlyJob( K3bAudioDoc* doc )
   : K3bBurnJob( )
 {
   m_doc = doc;
+  m_process = new KShellProcess();
 	
   m_iDocSize = doc->size();
   m_iTracksAlreadyWrittenSize = 0;
@@ -53,24 +55,12 @@ K3bAudioOnTheFlyJob::K3bAudioOnTheFlyJob( K3bAudioDoc* doc )
   
   m_streamingTimer = new QTimer( this );
   connect( m_streamingTimer, SIGNAL(timeout()), this, SLOT(slotTryWritingToProcess()) );
-
-  m_waitingForLengthTimer = new QTimer( this );
-  connect( m_waitingForLengthTimer, SIGNAL(timeout()), this, SLOT(slotTryStart()) );
-
-  //  m_ringBuffer = 0;
-
-
-  // testing
-  //-----------------------
-//   m_testFile = new QFile( "/home/trueg/download/test_k3b_job.cd" );
-//   m_testFile->open( IO_WriteOnly );
-  //-----------------------
-
 }
 
 
 K3bAudioOnTheFlyJob::~K3bAudioOnTheFlyJob()
 {
+  delete m_process;
 }
 
 
@@ -210,7 +200,15 @@ void K3bAudioOnTheFlyJob::slotParseCdrdaoOutput( KProcess*, char* output, int le
 
 void K3bAudioOnTheFlyJob::cancel()
 {
-  if( error() == K3b::WORKING ) {
+  if( m_process->isRunning() ) {
+    m_process->disconnect();
+
+    // cancel the module!!!
+    if( m_currentProcessedTrack != 0 )
+      m_currentProcessedTrack->module()->cancel();
+
+    m_process->closeStdin();
+    m_process->kill();
 
     // we need to unlock the writer because cdrecord locked it while writing
     bool block = m_doc->burner()->block( false );
@@ -219,13 +217,6 @@ void K3bAudioOnTheFlyJob::cancel()
     //    else if( k3bMain()->eject() )
     // m_doc->burner()->eject();
 
-
-    // cancel the module!!!
-    if( m_currentProcessedTrack != 0 )
-      m_currentProcessedTrack->module()->cancel();
-
-    m_process.closeStdin();
-    m_process.kill();
 	
     // remove toc-file
     if( QFile::exists( m_tocFile ) ) {
@@ -235,9 +226,8 @@ void K3bAudioOnTheFlyJob::cancel()
     }
   }
 
-  m_error = K3b::CANCELED;
   emit infoMessage( i18n("Writing canceled."), K3bJob::ERROR );
-  emit finished( this );
+  emit finished( false );
 }
 
 
@@ -248,34 +238,51 @@ void K3bAudioOnTheFlyJob::start()
 	
   emit started();
 	
-  m_error = K3b::WORKING;
-
   emit infoMessage( i18n("Waiting for all tracks' length to be calculated."), K3bJob::STATUS );
   emit infoMessage( i18n("This might take a while on slower systems."), K3bJob::STATUS );
   emit newSubTask( i18n("Preparing write process...") );
 
-  m_waitingForLengthTimer->start(100);
+  m_currentProcessedTrackNumber = 0;
+  m_currentProcessedTrack = m_doc->at(0);
+
+  slotRecalcNextTrackLength( true );
+}
+
+
+void K3bAudioOnTheFlyJob::slotRecalcNextTrackLength( bool success )
+{
+  if( !success ) {
+    emit infoMessage( i18n("Error while calculating length of track %1").arg(m_currentProcessedTrackNumber), K3bJob::ERROR );
+    emit finished( false );
+  }
+  else {
+    if( m_currentProcessedTrackNumber > 0 ) {
+      emit infoMessage( i18n("Length of track %1 calculated").arg(m_currentProcessedTrackNumber), K3bJob::STATUS );
+      m_currentProcessedTrack->module()->disconnect( this );
+      m_currentProcessedTrack = m_doc->at(m_currentProcessedTrackNumber);
+      m_currentProcessedTrackNumber++;
+    }
+    else
+      m_currentProcessedTrackNumber = 1;
+
+    if( m_currentProcessedTrack ) {
+      connect( m_currentProcessedTrack->module(), SIGNAL(finished(bool)), this, SLOT(slotRecalcNextTrackLength(bool)) );
+      m_currentProcessedTrack->module()->recalcLength();
+    }
+    else {
+      emit infoMessage( i18n("Finished length calculation"), K3bJob::STATUS );
+      slotTryStart();
+    }
+  }
 }
 
 
 void K3bAudioOnTheFlyJob::slotTryStart()
 {
-  bool ready = true;
-  for( K3bAudioTrack* track = m_doc->first(); track != 0; track = m_doc->next() ) {
-    if( !track->isAccurateLength() ) {
-      ready = false;
-      break;
-    }
-  }
-
-  if( ready ) {
-    m_waitingForLengthTimer->stop();
-
-    K3bEmptyDiscWaiter* waiter = new K3bEmptyDiscWaiter( m_doc->burner(), k3bMain() );
-    connect( waiter, SIGNAL(discReady()), this, SLOT(slotStartWriting()) );
-    connect( waiter, SIGNAL(canceled()), this, SLOT(cancel()) );
-    waiter->waitForEmptyDisc();
-  }
+  K3bEmptyDiscWaiter* waiter = new K3bEmptyDiscWaiter( m_doc->burner(), k3bMain() );
+  connect( waiter, SIGNAL(discReady()), this, SLOT(slotStartWriting()) );
+  connect( waiter, SIGNAL(canceled()), this, SLOT(cancel()) );
+  waiter->waitForEmptyDisc();
 }
 
 
@@ -285,8 +292,8 @@ void K3bAudioOnTheFlyJob::slotStartWriting()
 
   m_iDocSize = m_doc->size();
 	
-  m_process.clearArguments();
-  m_process.disconnect();
+  m_process->clearArguments();
+  m_process->disconnect();
 	
   firstTrack = true;
 	
@@ -307,65 +314,45 @@ void K3bAudioOnTheFlyJob::slotStartWriting()
     qDebug( "(K3bAudioJob) Could not write TOC-file." );
     emit infoMessage( i18n("Could not write correct TOC-file."), K3bJob::ERROR );
     
-    m_error = K3b::IO_ERROR;
-    emit finished( this );
+    emit finished( false );
   }
   else {
-    // initialize ring buffer with a minimal size of 4 000 000 bytes
-    // =======================================================================
-//     kapp->config()->setGroup("Audio Settings");
-//     m_bufferSize = kapp->config()->readNumEntry( "Buffer Size", 4000000 );
-//     if( m_bufferSize < 4000000 )
-//       m_bufferSize = 4000000;
-
-//     delete ringBuffer;
-//     m_ringBuffer = new char[m_bufferSize];
-//     m_bufferReader = m_bufferWriter = 0;
-    // =======================================================================
-
-    // ----------------
-    // now we need to fill the buffer and then start the writing from the buffer into the cdrdao-process
-    // ----------------
-
-
-
-
     kapp->config()->setGroup("External Programs");
-    m_process.clearArguments();
-    m_process << kapp->config()->readEntry( "cdrdao path" );
-    m_process << "write";
+    m_process->clearArguments();
+    *m_process << kapp->config()->readEntry( "cdrdao path" );
+    *m_process << "write";
 
     // device
-    m_process << "--device" << m_doc->burner()->genericDevice();
+    *m_process << "--device" << m_doc->burner()->genericDevice();
     if( m_doc->burner()->cdrdaoDriver() != "auto" ) {
-      m_process << "--driver";
+      *m_process << "--driver";
       if( m_doc->burner()->cdTextCapable() == 1 )
-	m_process << QString("%1:0x00000010").arg( m_doc->burner()->cdrdaoDriver() );
+	*m_process << QString("%1:0x00000010").arg( m_doc->burner()->cdrdaoDriver() );
       else
-	m_process << m_doc->burner()->cdrdaoDriver();
+	*m_process << m_doc->burner()->cdrdaoDriver();
     }
 			
     // additional parameters from config
     QStringList _params = kapp->config()->readListEntry( "cdrdao parameters" );
     for( QStringList::Iterator it = _params.begin(); it != _params.end(); ++it )
-      m_process << *it;
+      *m_process << *it;
     
     if( m_doc->dummy() )
-      m_process << "--simulate";
+      *m_process << "--simulate";
     if( k3bMain()->eject() )
-      m_process << "--eject";
+      *m_process << "--eject";
     
     // writing speed
-    m_process << "--speed" << QString::number(  m_doc->speed() );
+    *m_process << "--speed" << QString::number(  m_doc->speed() );
     
     // supress the 10 seconds gap to the writing
-    m_process << "-n";
+    *m_process << "-n";
     
     // toc-file
-    m_process << QString( "\"%1\"" ).arg(m_tocFile);
+    *m_process << QString("\"%1\"").arg(QFile::encodeName(m_tocFile));
     
     // debugging output
-    QStrList* _args = m_process.args();
+    QStrList* _args = m_process->args();
     QStrListIterator _it(*_args);
     while( _it ) {
       cout << *_it << " ";
@@ -374,18 +361,18 @@ void K3bAudioOnTheFlyJob::slotStartWriting()
     cout << endl << flush;
     
     // connect to the cdrdao slots
-    connect( &m_process, SIGNAL(processExited(KProcess*)),
+    connect( m_process, SIGNAL(processExited(KProcess*)),
 	     this, SLOT(slotCdrdaoFinished()) );
-    connect( &m_process, SIGNAL(receivedStderr(KProcess*, char*, int)),
+    connect( m_process, SIGNAL(receivedStderr(KProcess*, char*, int)),
 	     this, SLOT(slotParseCdrdaoOutput(KProcess*, char*, int)) );
-    connect( &m_process, SIGNAL(receivedStdout(KProcess*, char*, int)),
+    connect( m_process, SIGNAL(receivedStdout(KProcess*, char*, int)),
 	     this, SLOT(slotParseCdrdaoOutput(KProcess*, char*, int)) );
-    connect( &m_process, SIGNAL(wroteStdin(KProcess*)), 
+    connect( m_process, SIGNAL(wroteStdin(KProcess*)), 
 	     this, SLOT(slotWroteData()) );
 
     // first try to start the audiomodule
     m_currentProcessedTrackNumber = 0;
-    m_currentProcessedTrack = m_doc->first();
+    m_currentProcessedTrack = m_doc->at(0);
     m_currentProcessedTrack->module()->disconnect( this );
 
     connect( m_currentProcessedTrack->module(), SIGNAL(output(int)), 
@@ -394,25 +381,25 @@ void K3bAudioOnTheFlyJob::slotStartWriting()
 	     this, SLOT(slotModuleFinished(bool)) );
 
     if( m_currentProcessedTrack->module()->getStream() ) {
-      if( !m_process.start( KProcess::NotifyOnExit, KProcess::All ) )
+      if( !m_process->start( KProcess::NotifyOnExit, KProcess::All ) )
 	{
 	  // something went wrong when starting the program
 	  // it "should" be the executable
 	  qDebug("(K3bAudioOnTheFlyJob) could not start cdrdao");
-	  m_error = K3b::CDRDAO_ERROR;
 	  
+	  m_currentProcessedTrack->module()->cancel();
+
 	  // remove toc-file
 	  QFile::remove( m_tocFile );
 	  m_tocFile = QString::null;
 	  
 	  emit infoMessage( i18n("Could not start cdrdao!"), K3bJob::ERROR );
-	  emit finished( this );
+	  emit finished( false );
 	}
       else
 	{
 	  qDebug( "(K3bAudioOnTheFlyJob) process started!" );
 
-	  m_error = K3b::WORKING;
 	  if( m_doc->dummy() )
 	    emit infoMessage( i18n("Starting simulation at %1x speed...").arg(m_doc->speed()), K3bJob::STATUS );
 	  else
@@ -445,13 +432,13 @@ void K3bAudioOnTheFlyJob::slotWroteData()
 
 void K3bAudioOnTheFlyJob::slotTryWritingToProcess()
 {
-  if( m_currentModuleDataLength > 0 && m_process.isRunning() ) {
+  if( m_currentModuleDataLength > 0 && m_process->isRunning() ) {
     delete m_currentWrittenData;
     m_currentWrittenData = new char[m_currentModuleDataLength];
 
     int dataLength = m_currentProcessedTrack->module()->readData( m_currentWrittenData, m_currentModuleDataLength );
 
-    m_process.writeStdin( m_currentWrittenData, dataLength );
+    m_process->writeStdin( m_currentWrittenData, dataLength );
     
     m_currentModuleDataLength = 0;
     m_streamingStarted = true;
@@ -490,35 +477,35 @@ void K3bAudioOnTheFlyJob::slotModuleFinished( bool success )
       else {
 	// error
 	qDebug( "(K3bAudioOnTheFlyJob) Could not start streaming for track: " + m_currentProcessedTrack->fileName() );
-	m_process.closeStdin();
+	m_process->closeStdin();
       }
     }
     else {
       emit infoMessage( i18n("All tracks streamed"), K3bJob::STATUS );
       qDebug("(K3bAudioOnTheFlyJob) All tracks streamed. Closing cdrdao's Stdin."  );
-      m_process.closeStdin();
+      m_process->closeStdin();
     }
   }
   else {
     emit infoMessage( i18n("Error while streaming file"), K3bJob::ERROR );
-    m_process.closeStdin();
+    m_process->closeStdin();
   }
 }
 
 
 void K3bAudioOnTheFlyJob::slotCdrdaoFinished()
 {
-  if( m_process.normalExit() )
+  if( m_process->normalExit() )
     {
       // TODO: check the process' exitStatus()
-      switch( m_process.exitStatus() )
+      switch( m_process->exitStatus() )
 	{
 	case 0:
-	  m_error = K3b::SUCCESS;
 	  if( doc()->dummy() )
 	    emit infoMessage( i18n("Simulation successfully completed"), K3bJob::STATUS );
 	  else
 	    emit infoMessage( i18n("Writing successfully completed"), K3bJob::STATUS );
+	  emit finished( true );
 	  break;
 				
 	default:
@@ -526,24 +513,27 @@ void K3bAudioOnTheFlyJob::slotCdrdaoFinished()
 	  emit infoMessage( i18n("Cdrdao returned some error!"), K3bJob::ERROR );
 	  emit infoMessage( i18n("Sorry, no error handling yet! :-(("), K3bJob::ERROR );
 	  emit infoMessage( i18n("Please send me a mail with the last output..."), K3bJob::ERROR );
-	  m_error = K3b::CDRDAO_ERROR;
+	  m_currentProcessedTrack->module()->cancel();
+	  emit finished( false );
 	  break;
 	}
     }
   else
     {
-      m_error = K3b::CDRDAO_ERROR;
       emit infoMessage( i18n("Cdrdao did not exit cleanly!"), K3bJob::ERROR );
+      m_currentProcessedTrack->module()->cancel();
+      emit finished( false );
     }
 
   // remove toc-file
-  qDebug("(K3bAudioOnTheFlyJob) Removing temporary TOC-file");
-  QFile::remove( m_tocFile );
-  m_tocFile = QString::null;
-	
-  emit finished( this );
+  if( QFile::exists( m_tocFile ) ) {
+     qDebug("(K3bAudioOnTheFlyJob) Removing temporary TOC-file");
+     QFile::remove( m_tocFile );
+  }
 
-  m_process.disconnect();
+  m_tocFile = QString::null;
+
+  m_process->disconnect();
 }
